[level 2] Title: 피보나치 수, Time: 3.23 ms, Memory: 78.1 MB -BaekjoonHub

+++ 프로그래머스/lv2/12945. 피보나치 수/피보나치 수.java
@@ -1,5 +1,5 @@
 class Solution {
-    public double solution(int n) {
+    public int solution(int n) {
         double answer = 0;
         double n1 = 0, n2 = 1;
         int mod = 1234567;
